move JSON utilities into new class
